Update patch set 7
Patch Set 7:

> Hi Alon,
 > 
 > I cycled out with GENIUS folks here..
 > 
 > I understand that while we pop vlan and use a specific LPortTag
 > based on that and then drive the traffic,  it seems that the
 > general direction in GENIUS is to allow a given Tunnel Port to be
 > bound by multiple services and the respective service tables will
 > deal with how to decode packets coming off / going into such
 > tunnels.
 > 
 > Today we already have the mechanism to bind multiple services onto
 > an external tunnel port, except that it is not consistently driven
 > like how the VM endpoint ports are hooked on by services via the
 > LPORT_DISPATCHER_TABLE.
 > 
 > So it looks like the LPortDispatcher table will be the service
 > demultiplexing table rather than Classifier Table 0.
 > 
 > I am ok with either approach, as our focus is more on getting EVPN
 > to solve VXLAN based Inter-DC routing.
 > 
 > Since GENIUS already allows multiple services to bind on external
 > tunnel ports with all the services sharing the same LPortTag, we
 > were thinking better to stick with it..
 > 
 > We can always change to use Table 0 as service level classifier
 > like how VLAN Provider networks operate using a subsequent patch.
 > 
 > So the Spec and initial implementation we will use the
 > L3VNI_EXTERNAL_TUNNEL_DEMUX_TABLE for L3Forwarding and LPortTag for
 > external VXLAN Tunnel will be shared across services both L3VPN and
 > ELAN.

I'm mostly confused. I thought L3VNI_EXT_TUNNEL_DEMUX is only for external network usecases, not part of this spec?
Where do you propose to do the match on VNI and setting of lport/metadata? Table 17?

My only concerns are:

a. Keep provider networks behavior symmetric where possible. If it seems more correct to do it out of Genius, then so be it, but let's make sure we're deciding on the correct approach instead of the easy one. Whatever we decide should hold all provider networks if possible (VLAN, Flat, VXLAN), so if this stands we might need to consider moving provider network ports "demuxing".
I'd like Tali's input on this.

b. Simplify the pipeline as much as possible. It is already a complicated pipeline, having shared behaviors in the same table and not having extra tables like internal-tunnel table if it makes sense is great.

Patch-set: 7