Update patch set 1
Patch Set 1:

(2 comments)

This commit mostly refactors MDSAL operations to use MDSALUtils instead of some duplications.
As Tali noted even that is already outdated, and SingleTransactionDataBroker should be used. Maybe these refactorings should be done in a separate commit, and leave the logic itself for this commit?
It would be easier to merge the refactorings alone.

Patch-set: 1
Label: Code-Review=0