Update patch set 7
Patch Set 7:

>> I thought L3VNI_EXT_TUNNEL_DEMUX is only for external network usecases, not part of this spec?

YES L3VNI_EXT_TUNNEL_DEMUX is for VXLAN external network use-case only.
Mentioning and handling of the same will be retained as part of the EVPN_RT5 Spec only and not here.

>> Where do you propose to do the match on VNI and setting of lport/metadata? Table 17?

We are talking about INTERNAL VXLAN TUNNEL here and so where we decode packets should be decided in THIS SPEC .

Currently all VXLAN packets from INTERNAL VXLAN TUNNELS are decoded in the INTERNAL_TUNNEL_TABLE (Table 36).  This behaviour will be retained in this Spec.  The VNI matching will continue to be in INTERNAL_TUNNEL_TABLE and that matching will decide where further the packet goes into the ELAN PIPELINE.

This whole problem of comparing with VLAN and VXLAN packet handling is because of the way VLAN frames are handled different from VXLAN packets in the current pipeline. This spec does not attempt to solve that problem as that will be best solved by GENIUS which owns ITM and InterfaceManager.

Let the GENIUS spec for Tunnel Bindings come up for our review and then we can decide how to alter the pipelines to save the tables/flows based on for both INTERNAL and EXTERNAL tunnels.

>>My only concerns are:
>>a. Keep provider networks behavior symmetric where possible. If it seems >>more correct to do it out of Genius, then so be it, but let's make sure >>we're deciding on the correct approach instead of the easy one. Whatever >>we decide should hold all provider networks if possible (VLAN, Flat, >>VXLAN), so if this stands we might need to consider moving provider >>network ports "demuxing".

We donot fall on which is easy/difficult to implement as we are at the Spec level and haven't started anything on coding front.

As I said let GENIUS come up with consistent ways to handle tunnel ports 
using the tunnels-service-binding feature and then we can always patch this spec and implementation for consistency of both VLAN and VXLAN packets handling..

>>b. Simplify the pipeline as much as possible. It is already a >>complicated pipeline, having shared behaviors in the same table and not >>having extra tables like internal-tunnel table if it makes sense is >>great.

Completely agree here.  Am only saying we will revisit this Spec based on how GENIUS bring up the binding functionality.   Let us collaborate to make quick progress on getting VNIs being enforced as the focus area for this Spec.

Vivek

Patch-set: 7