Update patch set 1
Patch Set 1:

Hi Tali/Guy,

"after cluster reboot there is an attempt to handle ordering issues while DS loads on a case-to-case basis. I was wondering if there is or should be a plan for more robust cluster recovery implementation for Carbon" >> we would love that!
Most of our cluster reboot issues have the root cause in no control over ordering of DCNs, forcing us to handle it on a case to case basis, at times forcing us to place retries on reads.

Here elanTag is not written before a read on elanInstance is made at times, a retry makes it successful more often.

An improved fix on fetching elantag and writing here itself will soon be raised by Suraj, which eliminates the retry here.
But otherwise, this review should be okay to merge as it's harmless and helps take care of other listed improvements too..

Patch-set: 1