Table 21 ecmp flow not updated correctly
Description: When BGP neigbhor is added, quagga advertises routes towards
that DC-GW. If GRE tunnel towards that dc-gw is not configured, table 21
flows are not updated correctly.

This scenario was handled before but the code to handle this scenario
was removed somehow from sfi_oxygen branch. Copied missing code from REL61
branch and refactored to make it work.

Signed-off-by: Amitesh Soni <amitesh.soni@ericsson.com>
Change-Id: I6f54f7d4005dde7f2d05f53a05a27bf2b34a8cbb