Moving vpn-instance yang from VPNMgr > NeutronVPN
Vpn-instance and Vpn-interface Config Datastore are predominantly
used by the neutron-VPN layer, but the datastore is defined in
VPN manager. This review will take care of moving Vpn-instance
and associated yang models to neutron-VPN layer.

The second part of this change[1] will take care of moving
VPN-interface DS to neutronvpn

[1] https://git.opendaylight.org/gerrit/#/c/netvirt/+/87445/

Change-Id: I3c191d4f4acfc304548b0bfe954f80e75212d599
Signed-off-by: Ashik Alias <ashik.alias@ericsson.com>