Moving vpn-interface yang from VPNMgr > NeutronVPN
Vpn-instance and Vpn-interface Config Datastore are predominantly
used by the neutron-VPN layer, but the datastore is defined
in VPN manager.
This review will take care of moving Vpn-interface
and associated yang models to neutron-VPN layer.

The second part of this change[1] will take care of moving
VPN-instance DS to neutronvpn

[1] https://git.opendaylight.org/gerrit/#/c/netvirt/+/87418

Change-Id: Iad6e7a8508c5986bd12b8ef6d1e2d8294ecc02d9
Signed-off-by: Ashik Alias <ashik.alias@ericsson.com>