Id Manager allocateId optimization
Id Manager allocateId is called on all nodes for each of the SGs. This was
done for storing id in cache on all the nodes. This was resulting in
LockManager exception in IdManager. Though the exceptions were non
alarming, it was filling up karaf logs. Also, it was invoked on all the
nodes instread on entity owner node.

Now, allocateId for SG is moved to neutronvpn do that it gets executed
only on shard leader node. The id is then stored in Acl datastore for subsequently
to be used by ACL. In ACL, the ACL event listener gets it from Acl and
stores it in cache as it was done earlier.

Change-Id: I0b4c547ff2f4105523746069163cf6255c0cffff
Signed-off-by: Shashidhar Raja <shashidharr@altencalsoftlabs.com>