Stale MIP FIB/Flow entries present upon deletion of VRRP master
Issue:
======
This issue is seen when ports are pre-created and VM is booted on that.
In such condition , when VM is deleted , only south bound Interface remove
is received.
When VM which is a VRRP master(holding MIP-IP) was deleted , MIP-IP adj
was not deleted from the config-VpnInterface. Thus leading to issues like
duplicate FIB entry , when VM is booted back.

Solution:
=========
MIP-IP adj delete is handled appropriately when VM is deleted.

Signed-off-by: Karthikeyan Krishnan <karthikeyangceb007@gmail.com>
Change-Id: Ie42ff0eede90433c15b09537b7973842345ab376