ODL Service is not coming up
Issue:
======
There is a cyclic dependecny between Neutron Services and VPNManager
Services.
INeutronVpnManager service is exposed as part of NeutronManager.
ICentralizedSwitchProvider and IVpnManager services are exposed as part of
VPNManager.
Upon cluster reboot one of the node is not coming up properly as it is
failing
to start all bundle in active state.
dhcpservice-impl , natservice-impl , qosservice-impl ,statemanager-impl
are in GracePeriod
state because they are not able to resolve dependecny of
INeutronVpnManager .
INeutronVpnManager service is part of bundle neutronvpn-impl bundle which
is also in
GracePeriod state as it couldn't resolve dependecny
ICentralizedSwitchProvider
and IVpnManager.
ICentralizedSwitchProvider and IVpnManager are part of VPNManager which is
in Active state
but it failed to initialize the above mentioned service.
As it is not clear why VPNManager should not be able to expose the service
as mentioned
the only conclusion i can draw is the cyclic dependecny of
1. INeutronVpnManager service expose by NeutronManager in VPNManager
2. ICentralizedSwitchProvider , IVpnManager service expose by VPNManager
in NeutronManager
is cauing the failure in initializing the service in VPNManager.
The only reason VPNManager bundle is in active start is because its
service dependecny
for INeutronVpnManager is marked as Optional
With this conclusion this review remove this cyclic dependecny between
thease two modules.

Solution:
========
Removed neutronvpn dependency in vpnmanager

Signed-off-by: Karthikeyan Krishnan <karthikeyangceb007@gmail.com>
Change-Id: I6a3b2ce4dc84244267672d22b6e7ad124d280923